refactor: redesign the render crate into widgets crate
