+++ crates/filetype/Cargo.toml
@@ -6,7 +6,7 @@ edition.workspace = true
 
 [dependencies]
 shared.workspace = true
-render.workspace = true
+widgets.workspace = true
 config.workspace = true
 
 anyhow.workspace = true
