--- crates/render/README.md
@@ -1,4 +1,0 @@
-# About
-
-The `render` rust crate which maded for **Noti** notification system.
-This crate provides way to render banners or something other using widgets.
