--- crates/render/src/lib.rs
@@ -1,5 +1,0 @@
-pub mod color;
-pub mod drawer;
-pub mod image;
-pub mod types;
-pub mod widget;
