+++ crates/widgets/README.md
@@ -1,0 +1,4 @@
+# About
+
+The `widgets` crate provides UI components that can be used for composition, drawing and managing.
+It handles complex management providing nice API to use.
